+++ src/app/components/Modals/CreateEvents/CreateEventDetailModal.js
